BAU: Extract state machine execution
