+++ src/components/account-not-found/account-not-found-controller.ts
@@ -6,8 +6,8 @@ import { sendNotificationService } from "../common/send-notification/send-notifi
 import { BadRequestError } from "../../utils/error.js";
 import {
   getErrorPathByCode,
-  getNextPathAndUpdateJourney,
 } from "../common/constants.js";
+import { getNextPathAndUpdateJourney } from "../common/state-machine/state-machine-executor.js";
 import { USER_JOURNEY_EVENTS } from "../common/state-machine/state-machine.js";
 import xss from "xss";
 import { getServiceSignInLink } from "../../config.js";
