+++ src/components/account-recovery/change-security-codes-confirmation/change-security-codes-confirmation-controller.ts
@@ -2,7 +2,7 @@ import type { Request, Response } from "express";
 import { MFA_METHOD_TYPE } from "../../../app.constants.js";
 import type { ExpressRouteFunc } from "../../../types.js";
 import { USER_JOURNEY_EVENTS } from "../../common/state-machine/state-machine.js";
-import { getNextPathAndUpdateJourney } from "../../common/constants.js";
+import { getNextPathAndUpdateJourney } from "src/components/common/state-machine/state-machine-executor.js";
 import { getDefaultSmsMfaMethod } from "../../../utils/mfa.js";
 
 export function changeSecurityCodesConfirmationGet(): ExpressRouteFunc {
