+++ src/components/common/constants.ts
@@ -1,7 +1,5 @@
 import { PATH_NAMES } from "../../app.constants.js";
-import type { AuthStateContext } from "./state-machine/state-machine.js";
-import { getNextState } from "./state-machine/state-machine.js";
-import type { Request, Response } from "express";
+import type { Request } from "express";
 
 export const SECURITY_CODE_ERROR = "actionType";
 
@@ -180,39 +178,6 @@ export async function saveSessionState(req: Request): Promise<void> {
       }
     });
   });
-}
-
-export async function getNextPathAndUpdateJourney(
-  req: Request,
-  res: Response,
-  event: string,
-  ctx?: AuthStateContext,
-): Promise<string> {
-  const sessionId = res.locals.sessionId;
-
-  const nextState = getNextState(req.path, event, ctx);
-
-  req.session.user.journey = {
-    nextPath: nextState.value,
-    optionalPaths:
-      Object.keys(nextState.meta).length > 0
-        ? nextState.meta["AUTH." + nextState.value].optionalPaths
-        : [],
-  };
-
-  await saveSessionState(req);
-
-  req.log.info(
-    `User journey transitioned from ${req.path} to ${nextState.value} with session id ${sessionId}`
-  );
-
-  if (!nextState) {
-    throw Error(
-      `Invalid user journey. No transition found from ${req.path} with event ${event} with sessionId ${sessionId}`
-    );
-  }
-
-  return nextState.value;
 }
 
 export const JOURNEY_TYPE = {
