+++ src/components/common/mfa/send-mfa-controller.ts
@@ -8,8 +8,8 @@ import type { MfaServiceInterface } from "./types.js";
 import {
   ERROR_CODES,
   getErrorPathByCode,
-  getNextPathAndUpdateJourney,
 } from "../constants.js";
+import { getNextPathAndUpdateJourney } from "../state-machine/state-machine-executor.js";
 import { BadRequestError } from "../../../utils/error.js";
 import { USER_JOURNEY_EVENTS } from "../state-machine/state-machine.js";
 import { PATH_NAMES } from "../../../app.constants.js";
