+++ src/components/mfa-reset-with-ipv/mfa-reset-with-ipv-controller.ts
@@ -3,7 +3,7 @@ import { mfaResetAuthorizeService } from "./mfa-reset-authorize-service.js";
 import type { ExpressRouteFunc } from "../../types.js";
 import type { Request, Response } from "express";
 import { BadRequestError } from "../../utils/error.js";
-import { getNextPathAndUpdateJourney } from "../common/constants.js";
+import { getNextPathAndUpdateJourney } from "../common/state-machine/state-machine-executor.js";
 import { USER_JOURNEY_EVENTS } from "../common/state-machine/state-machine.js";
 import { MFA_METHOD_TYPE, PATH_NAMES } from "../../app.constants.js";
 export function mfaResetWithIpvGet(
