+++ src/components/reset-password-2fa-auth-app/reset-password-2fa-auth-app-controller.ts
@@ -3,8 +3,8 @@ import type { ExpressRouteFunc } from "src/types.js";
 import {
   ERROR_CODES,
   getErrorPathByCode,
-  getNextPathAndUpdateJourney,
 } from "../common/constants.js";
+import { getNextPathAndUpdateJourney } from "../common/state-machine/state-machine-executor.js";
 import { USER_JOURNEY_EVENTS } from "../common/state-machine/state-machine.js";
 import type { VerifyMfaCodeInterface } from "../enter-authenticator-app-code/types.js";
 import { verifyMfaCodeService } from "../common/verify-mfa-code/verify-mfa-code-service.js";
