+++ src/components/reset-password/reset-password-controller.ts
@@ -8,8 +8,8 @@ import {
 } from "../../utils/validation.js";
 import {
   ERROR_CODES,
-  getNextPathAndUpdateJourney,
 } from "../common/constants.js";
+import { getNextPathAndUpdateJourney } from "../common/state-machine/state-machine-executor.js";
 import { USER_JOURNEY_EVENTS } from "../common/state-machine/state-machine.js";
 import { BadRequestError } from "../../utils/error.js";
 import type { EnterPasswordServiceInterface } from "../enter-password/types.js";
