+++ src/components/setup-authenticator-app/setup-authenticator-app-controller.ts
@@ -3,8 +3,8 @@ import QRCode from "qrcode";
 import type { ExpressRouteFunc } from "../../types.js";
 import {
   ERROR_CODES,
-  getNextPathAndUpdateJourney,
 } from "../common/constants.js";
+import { getNextPathAndUpdateJourney } from "../common/state-machine/state-machine-executor.js";
 import { USER_JOURNEY_EVENTS } from "../common/state-machine/state-machine.js";
 import {
   generateQRCodeValue,
