+++ src/middleware/account-interventions-middleware.ts
@@ -1,5 +1,5 @@
 import type { NextFunction, Request, Response } from "express";
-import { getNextPathAndUpdateJourney } from "../components/common/constants.js";
+import { getNextPathAndUpdateJourney } from "src/components/common/state-machine/state-machine-executor.js";
 import { USER_JOURNEY_EVENTS } from "../components/common/state-machine/state-machine.js";
 import { accountInterventionService } from "../components/account-intervention/account-intervention-service.js";
 import type { ExpressRouteFunc } from "../types.js";
